+++ app/src/main/AndroidManifest.xml
@@ -25,8 +25,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.productDetail.ProductDetailActivity" />
-        <activity android:name=".ui.shoppingCart.ShoppingCartActivity" />
+        <activity android:name=".ui.product.detail.ProductDetailActivity" />
+        <activity android:name=".ui.cart.CartActivity" />
     </application>
 
 </manifest>
