+++ app/src/main/java/nextstep/shoppingcart/ui/product/list/component/ProductListTopBar.kt
@@ -1,9 +1,9 @@
-package nextstep.shoppingcart.ui.shoppingCart.component
+package nextstep.shoppingcart.ui.product.list.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material.icons.filled.ShoppingCart
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -19,7 +19,7 @@ import nextstep.shoppingcart.R
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun ShoppingCartTopBar(modifier: Modifier = Modifier, onClickBackIcon: () -> Unit) {
+fun ProductListTopBar(modifier: Modifier = Modifier, onClickCartIcon: () -> Unit) {
     TopAppBar(
         modifier = modifier
             .background(Color.White)
@@ -27,15 +27,15 @@ fun ShoppingCartTopBar(modifier: Modifier = Modifier, onClickBackIcon: () -> Uni
         title = {
             Text(
                 modifier = modifier.fillMaxWidth(),
-                text = stringResource(id = R.string.text_shopping_cart_title),
-                textAlign = TextAlign.Left
+                text = stringResource(id = R.string.text_product_list_title),
+                textAlign = TextAlign.Center
             )
         },
-        navigationIcon = {
-            IconButton(onClickBackIcon) {
+        actions = {
+            IconButton(onClick = onClickCartIcon) {
                 Icon(
-                    imageVector = Icons.Filled.ArrowBack,
-                    contentDescription = stringResource(id = R.string.text_back_Icon_description)
+                    imageVector = Icons.Filled.ShoppingCart,
+                    contentDescription = stringResource(id = R.string.text_shopping_cart_Icon_description)
                 )
             }
         }
@@ -44,6 +44,6 @@ fun ShoppingCartTopBar(modifier: Modifier = Modifier, onClickBackIcon: () -> Uni
 
 @Preview
 @Composable
-private fun ShoppingCartTopBarPreview() {
-    ShoppingCartTopBar(Modifier, {})
+private fun ProductListTopBarPreview() {
+    ProductListTopBar(Modifier, {})
 }
